pipeline_fetching/requirements.txt
typer==0.12.3
requests==2.32.3
tenacity==9.0.0
azure-storage-blob==12.20.0
python-dotenv==1.0.1
